1. 회원가입 시, City, Street, Zipcode 입력하도록 수정 2. 특정 필드만 update를 하도록, 엔티티에 @DynamicUpdate 적용 3. 팀 데이터의 STREET, CITY, ZIPCODE 필드 초기화

+++ src/main/java/com/side/portfolio/demo/controller/CartController.java
@@ -74,10 +74,10 @@ public class CartController {
             List<Cart> carts = cartService.findByTeamId(teamId);
             model.addAttribute("carts", carts);
 
-            Integer totalPrice = carts
+            Float totalPrice = carts
                     .stream()
                     .map(Cart::getPrice)
-                    .reduce(0, Integer::sum);
+                    .reduce((float) 0, Float::sum);
 
             Integer totalQty = carts
                     .stream()
@@ -107,7 +107,7 @@ public class CartController {
             Cart cart = Cart.builder()
                     .team(teamService.findById(form.getTeamId()))
                     .item(itemService.findById(form.getItemId()))
-                    .price(form.getPrice() * form.getQty())
+                    .price(Float.valueOf(form.getPrice()) * form.getQty())
                     .qty(form.getQty())
                     .createdDate(LocalDateTime.now())
                     .modifiedDate(LocalDateTime.now())
