+++ src/main/java/com/side/portfolio/demo/domain/Cart.java
@@ -26,13 +26,13 @@ public class Cart {
     private int qty;
 
     @Column(name = "cart_price")
-    private int price;
+    private float price;
 
     private LocalDateTime createdDate;
     private LocalDateTime modifiedDate;
 
     //정적 팩토리 메서드
-    public static Cart makeCart(Item item, int itemPrice, int qty) {
+    public static Cart makeCart(Item item, float itemPrice, int qty) {
         Cart cart = new Cart();
         cart.setItem(item);
         cart.setPrice(itemPrice);
@@ -41,7 +41,7 @@ public class Cart {
     }
 
     @Builder
-    public Cart(Team team, Item item, int price, int qty,
+    public Cart(Team team, Item item, float price, int qty,
                 LocalDateTime createdDate, LocalDateTime modifiedDate) {
         this.team = team;
         this.item = item;
